server_apps/user_request_handle/job-searcher_exact-match/requirements.txt
kafka-python==2.0.1
pymongo==3.10.1